requirements.txt
python-telegram-bot==20.7
flask==2.3.3
psycopg2-binary==2.9.9
pandas==2.2.1
python-dotenv==1.0.1
python-telegram-bot[job-queue]
setuptools>=70.0.0
